+++ entry-generation/godot-compiler-native-plugin/src/main/resources/META-INF/services/org.jetbrains.kotlin.compiler.plugin.CommandLineProcessor
@@ -1,1 +1,1 @@
-godot.compilerplugin.NativeGodotKotlinCompilerPluginCommandLineProcessor
+godot.compiler.plugin.NativeGodotKotlinCompilerPluginCommandLineProcessor
