+++ entry-generation/godot-compiler-native-plugin/src/main/resources/META-INF/services/org.jetbrains.kotlin.compiler.plugin.ComponentRegistrar
@@ -1,1 +1,1 @@
-godot.compilerplugin.NativeComponentRegistrar
+godot.compiler.plugin.NativeComponentRegistrar
