+++ entry-generation/godot-compiler-plugin-common/src/main/kotlin/godot/compiler/plugin/CompilerPluginConst.kt
@@ -1,4 +1,4 @@
-package godot.compilerplugin
+package godot.compiler.plugin
 
 import org.jetbrains.kotlin.config.CompilerConfigurationKey
