+++ entry-generation/godot-compiler-plugin/src/main/resources/META-INF/services/org.jetbrains.kotlin.compiler.plugin.CommandLineProcessor
@@ -1,1 +1,1 @@
-godot.compilerplugin.CommonGodotKotlinCompilerPluginCommandLineProcessor
+godot.compiler.plugin.CommonGodotKotlinCompilerPluginCommandLineProcessor
