+++ entry-generation/godot-compiler-plugin/src/main/resources/META-INF/services/org.jetbrains.kotlin.compiler.plugin.ComponentRegistrar
@@ -1,1 +1,1 @@
-godot.compilerplugin.CommonComponentRegistrar
+godot.compiler.plugin.CommonComponentRegistrar
